Move test_registry to tests only
