+++ Cargo.toml
@@ -48,7 +48,7 @@ serde = { version = "1", features = ["derive"] }
 serde_typename = "0.1"
 tar = "0.4"
 thiserror = "1.0.49"
-tokio = { version = "^1.26", features = ["fs", "rt", "macros", "process", "io-std", "tracing", "rt-multi-thread"] }
+tokio = { version = "^1.26", features = ["fs", "rt", "macros", "process", "io-std", "tracing"] }
 toml = "0.8.0"
 tonic-build = { version = "0.10.0", optional = true }
 tracing = "0.1"
@@ -56,11 +56,11 @@ tracing-subscriber = "0.3"
 url = { version = "2.4", features = ["serde"] }
 walkdir = "2"
 sha2 = "0.10.8"
-axum = { version = "0.7.2", default-features = false, features = ["tokio", "http1"] }
 
 [dev-dependencies]
 assert_cmd = "2.0"
 assert_fs = "1.0"
+axum = { version = "0.7.2", default-features = false, features = ["tokio", "http1"] }
 fs_extra = "1.3"
 hex = "0.4.3"
 paste = "1.0.14"
@@ -69,6 +69,7 @@ pretty_assertions = "1.4"
 serde_json = { version = "1.0.107" }
 serde_test = "1.0.176"
 similar-asserts = "1.5.0"
+tokio = { version = "^1.26", features = ["rt-multi-thread"] }
 
 [profile.deb]
 inherits = "release"
